ENGINE: ELF Building and Reading

+++ src/commonMain/kotlin/cengine/lang/asm/run/AsmExecutable.kt
@@ -1,9 +1,9 @@
 package cengine.lang.asm.run
 
 import cengine.editor.annotation.Severity
+import cengine.lang.LanguageService
 import cengine.lang.RunConfiguration
 import cengine.lang.asm.AsmLang
-import cengine.lang.asm.ast.TargetSpec
 import cengine.lang.asm.ast.impl.ASNode
 import cengine.lang.asm.elf.ELFBuilder
 import cengine.lang.asm.elf.Ehdr
@@ -14,23 +14,23 @@ import emulator.kit.nativeError
 import emulator.kit.nativeInfo
 import emulator.kit.nativeWarn
 
-class AsmExecutable(val asmLang: AsmLang) : RunConfiguration.FileRun {
+class AsmExecutable() : RunConfiguration.FileRun<LanguageService> {
     companion object {
         const val EXECUTABLE_SUB_DIR = "exec"
     }
 
     override val name: String = "ELF Executable"
-    private val spec: TargetSpec get() = asmLang.spec
 
-    override fun run(file: VirtualFile, vfs: VFileSystem) {
-        val asmFile = asmLang.psiParser.parseFile(file, null)
+    override fun run(file: VirtualFile, lang: LanguageService, vfs: VFileSystem) {
+        if (lang !is AsmLang) return
+        val asmFile = lang.psiParser.parseFile(file, null)
 
-        val outputPath = "${AsmLang.OUTPUT_DIR}${VFileSystem.DELIMITER}$EXECUTABLE_SUB_DIR${VFileSystem.DELIMITER}${file.name.removeSuffix(asmLang.fileSuffix)}.o"
+        val outputPath = "${AsmLang.OUTPUT_DIR}${VFileSystem.DELIMITER}$EXECUTABLE_SUB_DIR${VFileSystem.DELIMITER}${file.name.removeSuffix(lang.fileSuffix)}.o"
 
         vfs.deleteFile(outputPath)
         val outputFile = vfs.createFile(outputPath)
 
-        val builder = ELFBuilder(spec.ei_class, spec.ei_data, spec.ei_osabi, spec.ei_abiversion, Ehdr.ET_EXEC, spec.e_machine)
+        val builder = ELFBuilder(lang.spec, Ehdr.ET_EXEC)
         val content = builder.build(*asmFile.children.filterIsInstance<ASNode.Statement>().toTypedArray())
 
         val collector = PsiNotationCollector()
@@ -44,7 +44,7 @@ class AsmExecutable(val asmLang: AsmLang) : RunConfiguration.FileRun {
             }
         }
 
-        if (collector.notations.filter { it.severity == Severity.ERROR }.isEmpty()) {
+        if (collector.notations.none { it.severity == Severity.ERROR }) {
             outputFile.setContent(content)
         }
     }
